feat: 외부에서 데이터를 받아오도록 컴포넌트 수정
